Add project scaffolding
Add a minimal C program. Add a cmake file with custom targets to flash
and connect GDB in addition to the targets imported from the Raspberry
Pi Pico SDK.

The cmake assumes the following environment variables are set:

  - PICO_SDK_PATH: path to the root of the Raspberry Pi Pico SDK [1]
  - OPENOCD_BIN_PATH: path to the openocd binary [2]
  - OPENOCD_SCRIPTS_PATH: path to openocd's scripts directory

The flash target assumes the Pico is connected to another Pico running
picoprobe.

See [3] for more information on how to get started.

1. https://github.com/raspberrypi/pico-sdk
2. https://github.com/raspberrypi/openocd
3. https://datasheets.raspberrypi.org/pico/getting-started-with-pico.pdf

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>