Add program states
Introduce states corresponding to what hardware button the user is
currently pressing. Transitions are only valid from the idle state to
one of the actives states, and immediately back again: if a second
button is pressed when already in a state, that button press is silently
ignored.

Also introduce timed transitions between three phases: the first,
short-lived one, has all LEDs on at full brightness. The second dims the
LEDs. The final one turns off the LEDs for the glow and start buttons,
and ignore any button presses: this prevents the user from accidentally
engaging the starter engine when the engine is (presumably) already
running.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>