Update make files to produce .uf2
The .uf2 artifacts can be used to flash a Pico without the need for an
additional Pico. While this requires manually resetting the Pico into
flash mode, it's the more convenient mode for a one-off update of the
Pico installed in the boat. To flash a .uf2 file:

  1. connect the Pico via USB cable while holding down the BOOTSEL
     button on the Pico
  2. release BOOTSEL, wait for /dev/disk/by-label/RPI-RP2 to appear
  3. mount the disk
  4. copy the .uf2 file to the mount point
  5. unmount the disk
  6. disconnect the Pico from USB

Note: this change does not interfere with the picobridge setup, which
can still be used.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>