Improve build files and README
Include a copy of pico-sdk's cmake file which supports using a
pre-installed version of the SDK or downloading it on the fly.

Also add default values to the OpenOCD variables and update the README
to include the Arch Linux AUR packages needed to make building as simple
as `cmake .. && make`.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>